backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-multipart
pydantic[email]
